requirements.txt
langchain
langchain_community
langchain_google_genai
langchain-google-vertexai
langchain-cohere
langchain-aws
langchain-openai
langchain-huggingface
langgraph
python-dotenv
pandas
tqdm
numpy
torch
transformers
flask
ipykernel
sentencepiece
bitsandbytes
PyPDF2
textract
selenium
faiss-cpu